requirements.txt
numpy
pandas
scikit-learn
pyyaml
joblib
mlflow
dvc
fastapi 
uvicorn